+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## develop branch
 
-Nothing yet.
+### New
+
+* Checks\IsObjectOfType - added
+* Requirements\RequireObjectOfType - added
 
 ## 2.13.0 - Sun Sep 20 2015